requirements/base.txt
psycopg2
Pillow
Flask
SQLAlchemy
Flask-SQLAlchemy
Flask-Admin
Flask-Environments
jinja2
requests
nose
python-dateutil
ipython

Flask-Security==1.7.4
Flask-Script==2.0.5
Flask-Login==0.2.11
